+++ src/app.html
@@ -19,7 +19,7 @@
 
 		<link rel="icon" href="%sveltekit.assets%/favicon.ico" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<title>Scraper | Billabong and RVCA</title>
+		<title>Netta's Automation | Scrape Billabong + RVCA</title>
 		%sveltekit.head%
 	</head>
 	<body data-sveltekit-preload-data="hover" data-theme="modern">
