+++ src/lib/ui/Header.svelte
@@ -1,0 +1,4 @@
+<section class="mb-3 text-center">
+	<h1 class="h1 mb-2">Netta's Automation 🚀</h1>
+	<p>Paste a list of models, and we will scrape the product pages for the materials</p>
+</section>
